refactor: Rename Option to Settings
