+++ src/components/Settings/BufferSelection.tsx
@@ -2,7 +2,7 @@ import { SettingsContext } from "@/context/SettingsContext";
 import "@/styles/table.css";
 import { Speffz } from "@/utils/types/Speffz";
 import { useContext } from "react";
-import styles from "./Option.module.css";
+import styles from "./Settings.module.css";
 
 function BufferSelectionBox({
   buffer,
